Rename mock file and update title and description mocks
