+++ codesign/api/__tests__/report/uploadReport-test.tsx
@@ -2,7 +2,7 @@ import { mockFetchSuccess, mockFetchError } from '@/mocks/mockFetch';
 import { ApiResponse, ReportUploadSuccess } from '@/types/api';
 import { ROUTES } from '@/constants/api/routes';
 import { uploadReport } from '@/api/report/uploadReport';
-import { createReportFormDetails } from '@/mocks/Report';
+import { createReportFormDetails } from '@/mocks/mockReport';
 import { ReportFormDetails } from '@/types/Report';
 
 describe('uploadReport', () => {
